Added a way to add your own evolutions within the gpu implementation. It currently always needs the original states. It's possible to have the following evolution pattern:
It changes from state i to j by default.
It changes from state i to j on a certain amount of neighbours
It changes from state i to j with a range of neighbours
It changes from state i to j when the amount of neighbours is less then
It changes from state i to j when the amount of neighbours is more then

When nothing is given for a certain i with it's neighbours, the state does not change.
This can happen in 2 ways: i is not given or nothing is defined for a that amount of neighbours.

If time allows, more ways will be implemented.